+++ CHC/src/main.py
@@ -5,13 +5,14 @@ import uuid
 import boto3
 from baseaws.shared_functions import ContainerServices
 import requests
+from requests.status_codes import codes as status_codes
 
 CONTAINER_NAME = "CHC"    # Name of the current container
 CONTAINER_VERSION = "v3.0"      # Version of the current container
 VIDEO_FORMATS = ['mp4','avi']
 IMAGE_FORMATS = ['jpeg','jpg','png']
 
-def request_processing(client, container_services, body):
+def request_processing(client, container_services, body)->bool:
     """Converts the message body to json format (for easier variable access)
     and sends an API request for the ivs feature chain container with the
     file downloaded to be processed.
@@ -59,10 +60,12 @@ def request_processing(client, container_services, body):
         response = requests.post(addr, files=files, data=payload)
         logging.info("API POST request sent! (uid: %s)", uid)
         logging.info("IVS Chain response: %s", response.text)
+        return response.status_code == status_codes.ok
     except requests.exceptions.ConnectionError:
         logging.info("\n######################## Exception #########################")
         logging.exception("The following exception occured during execution:")
         logging.info("############################################################\n")
+        return False
 
     # TODO: ADD EXCEPTION HANDLING IF API NOT AVAILABLE (except Exception as e:)
 
@@ -161,10 +164,11 @@ def main():
             # save some messages as examples for development
             #log_message(message)
             # Processing request
-            request_processing(s3_client,container_services,message['Body'])
+            success = request_processing(s3_client,container_services,message['Body'])
 
             # Delete message after processing
-            container_services.delete_message(sqs_client,
+            if success:
+                container_services.delete_message(sqs_client,
                                               message['ReceiptHandle'])
 
         # Check API SQS queue for new update messages
